chore: move DeepPartial type to internal util
